requirements.txt
# Core Web Framework
fastapi
uvicorn

# Database ORM and Tooling
sqlalchemy
alembic

# Data Validation and Configuration
pydantic
python-decouple

# AI & Vector DB Dependencies
chromadb
sentence-transformers
librosa>=0.9.2
tqdm

# Core ML Libraries (needed by sentence-transformers)
torch
torchvision
torchaudio
scikit-learn
lyricsgenius
sentencepiece
scikit-learn
musicbrainzngs
